Fix consistency of ETag: generation
While HTTP ETag should be opaque per RFC9110,
in this RFC there is a convention of generating ETag
like so:  Etag = offer-ice-ufrag + ':' + answer-ice-ufrag

Except, the values used in the first example are inconsistent
and confusing.

It might be argued that since ETag should be opaque,
this inconsistency doesn't matter.
But I find it confusing, and thinks others might also.
(In the patch examples, you can see how the ETag is generated on
an Ice restart)

It might also benefit the RFC to explain both:
1. The convention use for generating ETag
2. Why the convention isn't required (or is required) in the WHIP RFC